+++ src/Raven.CodeAnalysis/Macros/MethodMacroExecutorAdapter.cs
@@ -27,6 +27,7 @@ internal sealed class MethodMacroExecutorAdapter : IMacroExecutor
     public string? Documentation => _definition.Documentation;
     public DocumentationFormat DocumentationFormat => _definition.DocumentationFormat;
     public MacroInvocationTargets InvocationTargets => _definition.InvocationTargets;
+    public Type? ExpressionResultType => _definition.ExpressionResultType;
     public MacroCarrierKinds CarrierKinds => _definition.CarrierKinds;
     public MacroBodyRequirement BodyRequirement => _definition.BodyRequirement;
     public bool AcceptsArguments => Parameters.Any(static parameter => parameter.InvocationArgumentOrdinal is not null);